pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "project for learning agentic ai foundations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.3.0",
    "langchain-groq>=1.1.1",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.1.1",
    "pypdf>=6.4.2",
    "gradio>=6.1.0",
    "tavily>=1.1.0",
    "ipywidgets>=8.1.8",
    "sounddevice>=0.5.3",
    "scipy>=1.16.3",
    "langgraph-cli[inmem]>=0.4.11",
    "langchain-mcp-adapters>=0.2.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
